+++ packages/dev-deployment-quarkus-blank-app-image/README.md
@@ -18,7 +18,7 @@
 # Dev Deployment Quarkus Blank App Image
 
 This image is ready to be used for Dev Deployments on KIE Sandbox.
-It starts the dev-deployment-upload-service and then places the uploaded files inside a blank Kogito Quarkus app.
+It starts the dev-deployment-upload-service and then places the uploaded files inside a blank Quarkus Kogito app.
 These files can decisions or processes, all of them will be used as resources for the app.
 
 ## Build arguments
