+++ packages/online-editor/src/devDeployments/services/deploymentOptions/customImage/DeploymentYaml.ts
@@ -45,7 +45,7 @@ spec:
     spec:
       containers:
         - name: \${{ devDeployment.uniqueName }}
-          image: ${args.kogitoQuarkusBlankAppImageUrl}
+          image: ${args.quarkusBlankAppImageUrl}
           imagePullPolicy: ${args.imagePullPolicy}
           resources:
             limits:
