+++ packages/online-editor/src/devDeployments/services/deploymentOptions/quarkusBlankApp/DeploymentYaml.ts
@@ -48,7 +48,7 @@ spec:
           emptyDir: {}
       containers:
         - name: \${{ devDeployment.uniqueName }}
-          image: ${args.kogitoQuarkusBlankAppImageUrl}
+          image: ${args.quarkusBlankAppImageUrl}
           imagePullPolicy: ${args.imagePullPolicy}
           resources:
             limits:
