+++ packages/online-editor/src/devDeployments/services/deploymentOptions/quarkusBlankApp/IngressYaml.ts
