+++ packages/online-editor/src/devDeployments/services/deploymentOptions/quarkusBlankApp/RouteYaml.ts
