+++ packages/online-editor/src/devDeployments/services/deploymentOptions/quarkusBlankApp/ServiceYaml.ts
