+++ packages/online-editor/src/devDeployments/services/deploymentOptions/quarkusBlankApp/index.ts
@@ -22,7 +22,7 @@ import { DeploymentOption, DeploymentOptionOpts } from "../types";
 import { DeploymentYaml } from "./DeploymentYaml";
 import { ServiceYaml } from "./ServiceYaml";
 
-export function KogitoQuarkusBlankAppOption(args: DeploymentOptionArgs, opts?: DeploymentOptionOpts): DeploymentOption {
+export function QuarkusBlankAppOption(args: DeploymentOptionArgs, opts?: DeploymentOptionOpts): DeploymentOption {
   return {
     name: "Quarkus Blank App",
     content: `
